Rename packages and classes.
