+++ src/pro/sky/javadev/lesson3/Homework03ConditionalOperators.java
@@ -1,0 +1,8 @@
+package pro.sky.javadev.lesson3;
+
+public class Homework03ConditionalOperators {
+    public static void main(String[] args) {
+
+
+    }
+}
